% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weights.R
\name{calc_weights}
\alias{calc_weights}
\title{Calculate Precision Weights From Gamma Regression}
\usage{
calc_weights(data, identifier, design, gamma_reg_model)
}
\arguments{
\item{data}{A \code{data.frame} who's quantitative columns should be converted to
precision weights.}

\item{gamma_reg_model}{a \code{glm} object as produced by
see \code{\link[mavis]{fit_gamma_regression}} for details.}
}
\value{
The same \code{data.frame} but with all quantitative values replaced by
their precision weights.
}
\description{
Calculates the precision weights for each measurement.
}
\details{
\code{calc_weights} takes as input a data frame and a \code{\link[stats]{glm}}  object produced
by fit_gamma_regression,
see \code{\link[mavis]{fit_gamma_regression}} for details.
For all numeric columns, it predicts the standard deviation using the gamma
regression. It then squares and takes the reciprocal of each value to generate the
precision weights.
}
\examples{
# Generate a design matrix for the data
design <- model.matrix(~ 0 + factor(rep(1:2, each = 3)))

# Set correct colnames, this is important for fit_gamma_weights
colnames(design) <- paste0("ng", c(50, 100))

# Normalize and log-transform the data
yeast <- psrn(yeast_prog, "identifier") \%>\%
  # Add row means and variances
  calculate_mean_sd_trends(design)

# Fit the gamma regression model for the mean-variance trend
gamma_model <- fit_gamma_regression(yeast, sd ~ mean)

# Generate the weights for the yeast data
calc_weights(yeast, gamma_model)
}
