% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_priors.R
\name{estimate_gamma_hyperparameters}
\alias{estimate_gamma_hyperparameters}
\title{Estimate Gamma priors for sigma}
\usage{
estimate_gamma_hyperparameters(gamma_reg, data, design_matrix)
}
\arguments{
\item{gamma_reg}{A Gamma regression object}

\item{data}{A \code{tibble} or \code{data.frame} to add gamma priors to}

\item{design_matrix}{A design matrix for the data (see example)}
}
\value{
A \code{tibble} or \code{data.frame} with the alpha,beta priors
}
\description{
Estimates the priors for the Bayesian model.
}
\examples{
# Setup model matrix
design <- model.matrix(~ 0 + factor(rep(1:2, each = 3)))
colnames(design) <- paste0("ng", c(50, 100))

# Normalize data
yeast <- yeast \%>\%
    psrn("identifier") \%>\%
    # Get mean-variance trends
    calculate_mean_sd_trends(design)

# Fit gamma regression
gamma_model <- fit_gamma_regression(yeast, sd ~ mean)

# Estimate priors
gam_reg \%>\%
    estimate_gamma_hyperparameters(yeast, design)
}
