% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_uncertainty.R
\name{estimate_uncertainty}
\alias{estimate_uncertainty}
\title{Estimate measurement uncertainty}
\usage{
estimate_uncertainty(data, identifier, design_matrix, gam_reg)
}
\arguments{
\item{data}{A \code{tibble} or \code{data.frame}}

\item{identifier}{Id column}

\item{design_matrix}{Cell mean design matrix for the data}

\item{gam_reg}{A gamma regression model}
}
\value{
A matrix with the uncertainty
}
\description{
Estimates the measurement uncertainty for each data point using
a Gamma regression.
}
\examples{
# Please see vignette('baldur') for examples
}
