% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imp}
\alias{imp}
\alias{get_imputation_pars}
\alias{single_imputation}
\title{Imputation}
\usage{
get_imputation_pars(data, design, formula = sd_p ~ mean, workers = 1, ...)

single_imputation(
  data,
  design,
  formula = sd_p ~ mean,
  workers = 1,
  imp_pars = NULL,
  ...
)
}
\arguments{
\item{data}{a \code{data.frame} to perform the imputation on, missing values should
be \code{NA}.}

\item{design}{a design or model matrix as produced by
\code{\link[stats]{model.matrix}} with column names corresponding to the}

\item{formula}{Regression formula for the pooled variance and mean trend}

\item{workers}{Number of parallel workers to use when building the forest}

\item{...}{Additional arguments to \link[ranger:ranger]{ranger::ranger}.}

\item{imp_pars}{Imputation parameters to use; as inferred by \link{get_imputation_pars}.}
}
\value{
a \code{list} containing the imputation parameters and corresponding index.

a \code{data.frame} with \code{NA} values replaced by imputed values.
}
\description{
Imputation

Estimates the parameters for the imputation model.

Performs a single imputation run and returns the data with NA values replaced
by imputed values.
}
\examples{
# Generate a design matrix for the data
design <- model.matrix(~ 0 + factor(rep(1:2, each = 3)))

# Set correct colnames, this is important.
colnames(design) <- paste0("ng", c(50, 100))
\donttest{
# Estimate the imputation parameters
imp_pars <- yeast \%>\%
  psrn("identifier") \%>\%
  get_imputation_pars(design)
}
\donttest{
# Run the imputation
yeast \%>\%
  single_imputation(design, imp_pars = imp_pars)
}
}
