% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gamma_regressions.R
\name{Mean-Variance_Gamma_Regressions}
\alias{Mean-Variance_Gamma_Regressions}
\alias{fit_gamma_regressions}
\alias{fit_gamma_imputation}
\alias{fit_gamma_weights}
\title{Functions for Fitting the Mean-Variance Gamma Regression Models}
\usage{
fit_gamma_regressions(data, formula = sd ~ mean + c)

fit_gamma_imputation(data, design, id_col = "id")

fit_gamma_weights(data, design, id_col = "id")
}
\arguments{
\item{data}{a \code{data.frame} to generate the mean-variance trends for. It
should contain columns with conditions named as the column names in
\code{design} (presumably with some suffix).}

\item{design}{a design or model matrix as produced by
\code{\link[stats]{model.matrix}} with column names corresponding to the
different conditions.}

\item{id_col}{a character for the name of the column containing the
name of the features in data (e.g., peptides, proteins, etc.).}
}
\value{
\code{fit_gamma_imputation} returns a named
list where the names corresponds to the conditions. Each index contains
a \code{glm} object with the gamma regression for the mean-variance trend.
\code{fit_gamma_weights} returns a \code{glm} object with the gamma regression
for the precision weights. \code{fit_gamma_regressions} returns a list with
the results from \code{fit_gamma_imputation} in \verb{$imputation} and the results
from \code{fit_gamma_weights} in \verb{$weights}.

gamma regressions for each condition

gamma regression for weights
}
\description{
\code{fit_gamma_regressions} is a wrapper function that calls both
\code{fit_gamma_imputation} and
\code{fit_gamma_weights}. It returns a list containing
the models for imputation in \verb{$imputation} and the weights in \verb{$weights}.
\code{fit_gamma_imputation} returns a list named according to the different
conditions and \code{fit_gamma_weights} returns a \code{glm} object containing the
gamma regression for the mean-variance trend.
}
\section{Functions}{
\itemize{
\item \code{fit_gamma_regressions}: Wrapper function that runs both
\code{fit_gamma_imputation} and \code{fit_gamma_weights}

\item \code{fit_gamma_imputation}: Function that generates per
condition mean-variance trends used in the imputation procedure.
Each id in the \code{id_col} gets one mean and variance calculated for each
condition. One model is then fitted per condition.

\item \code{fit_gamma_weights}: Function to produce the
mean-variance trend used to calculate the precision weights used in
\code{\link[limma]{lmFit}}. Each id in the \code{id_col} gets one mean and one
variance across all conditions and one model is then fitted for all
mean-variance pairs.
}}

\examples{
# Generate a design matrix for the data
design <- model.matrix(~ 0 + factor(rep(1:2, each = 3)))

# Set correct colnames, this is important for fit_gamma_*
colnames(design) <- paste0("ng", c(50, 100))

# Normalize and log-transform the data
yeast <- psrn(yeast_prog, "identifier")

# Fit all gamma regression models for the mean-variance trends
all_gamma_models <- fit_gamma_regressions(yeast, design, "identifier")

# Fit the gamma regression models for the mean-variance trend used in the
# imputation procedure
gamma_imputation_models <- fit_gamma_imputation(yeast, design, "identifier")

# Fit the gamma regression model for the mean-variance trend used for
# estimating the precision weights used in limma
gamma_weight_model <- fit_gamma_weights(yeast, design, "identifier")
}
