% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{mean2}
\alias{mean2}
\title{Calculates the mean of the summed variable x}
\usage{
mean2(x, n)
}
\arguments{
\item{x}{Sum of variables to calculate the mean from}

\item{n}{Number of variables that were summed}
}
\value{
A numerical value for the mean of x
}
\description{
Calculates the mean of the summed variable x
}
\examples{
# Common
x <- 1
y <- 2
z <- add(x, y)
# Unique
n <- 2
mean2(z, n)
}
