% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_partitioning.R
\name{multi_trend_partitioning}
\alias{multi_trend_partitioning}
\title{Multiple Trend Partitioning}
\usage{
multi_trend_partitioning(
  data,
  design_matrix,
  formula = sd ~ mean + c,
  h = 0.1,
  verbose = 0,
  penalty = p,
  plot = TRUE
)
}
\arguments{
\item{data}{A \code{tibble} or \code{data.frame} to partition}

\item{design_matrix}{A design matrix for the data (see example)}

\item{formula}{Formula for the Gamma regression}

\item{h}{Size of the integration window}

\item{verbose}{Should the procedure be printed (1 for multi partitioning, 2
for multi and trend partitioning, 0 for no).}

\item{penalty}{Penalty function for the penalized likelihood, defaults to \link{p}}

\item{plot}{Should a plot of the final solution be printed?}
}
\value{
a \code{tibble} or \code{data.frame} with \eqn{\boldsymbol{\hat{C}}} added as a new column.
}
\description{
Performs multiple trend partitioning on the mean-variance trend and adds a
new column classifying each peptide into a cluster.
}
\examples{
# Normalize data
design <- model.matrix(~ 0 + factor(rep(1:2, each = 3)))
colnames(design) <- paste0("ng", c(50, 100))

yeast_norm <- yeast \%>\%
  psrn("identifier")
yeast_norm \%>\%
  calculate_mean_sd_trends(design) \%>\%
  multi_trend_partitioning(design)
}
