% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_partitioning.R
\name{p}
\alias{p}
\title{Penalty function}
\usage{
p(k, n)
}
\arguments{
\item{k}{The number of clusters}

\item{n}{The number of data points}
}
\value{
A real number of the penalty of the likelihood
}
\description{
Used to calculate and subtract a penalty from the likelihood.
}
\examples{
p(1, 10)
}
