% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ma.R
\name{plot_ma}
\alias{plot_ma}
\title{Generate a MA-plot of the analysis.}
\usage{
plot_ma(hits, data = NULL, id_col = NULL, alpha = NULL, abs_lfc = NULL)
}
\arguments{
\item{hits}{a \code{tibble} produced by \code{\link[mavis]{extract_results}}}

\item{data}{the \code{data.frame} used in the analysis.
This should always be \code{NULL} if \code{hits} was produced by
\code{\link[mavis]{extract_results}}.}

\item{id_col}{a character for the name of the column containing the
name of the features in data (e.g., peptides, proteins, etc.).
This should always be \code{NULL} if \code{hits} was produced by
\code{\link[mavis]{extract_results}}.}

\item{alpha}{The alpha cut-off for considering a p-value significant.
This should always be \code{NULL} if \code{hits} was produced by
\code{\link[mavis]{extract_results}}.}

\item{abs_lfc}{If a LFC threshold should also be used in the decision.
This should always be \code{NULL} if \code{hits} was produced by
\code{\link[mavis]{extract_results}}.}
}
\value{
a \code{ggplot2} of the distribution of the hits.
}
\description{
Generate a MA-plot of the analysis.
}
\examples{
# Generate a design matrix for the data
design <- model.matrix(~ 0 + factor(rep(1:2, each = 3)))

# Set correct colnames, this is important for fit_gamma_*
colnames(design) <- paste0("ng", c(50, 100))

# Generate the contrast matrix
contrast <- limma::makeContrasts(
  contrasts = "ng100-ng50",
  levels = design
)

# Normalize and log-transform the data
yeast <- psrn(yeast, "identifier")
\dontrun{

results <- run_pipeline(yeast, design, contrast, 1000, 5, "identifier", TRUE)
imputation_summary <- extract_results(yeast, results, .05, 1, "fdr", "identifier")
plot_ma(imputation_summary)
}
}
