% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_norm_box.R
\name{plot_norm_box}
\alias{plot_norm_box}
\title{Generates a box-plot for \code{log2} transformed raw, tmm-, and psn- normalized
values.}
\usage{
plot_norm_box(
  data,
  id_col = "id",
  trim_M = 0.3,
  trim_A = 0.05,
  norm_target = NULL,
  plot_target = NULL,
  reference_sample = NULL
)
}
\arguments{
\item{data}{data.frame containing the data to normalize}

\item{id_col}{a character for the name of the column containing the
name of the features in data (e.g., peptides, proteins, etc.)}

\item{trim_M}{percent of fold-change values to trim}

\item{trim_A}{percent of means to trim}

\item{norm_target}{target columns to normalize, supports
\code{\link[tidyselect]{tidyselect-package}} syntax. By default, all numerical
columns will be used in the normalization if not specified.}

\item{plot_target}{target columns to plot, supports
\code{\link[tidyselect]{tidyselect-package}} syntax. By default, all numerical
columns are plotted.}

\item{reference_sample}{Specify a reference sample to normalize to in the
\code{\link[mavis]{tmm}} method}
}
\value{
ggplot with samples on the x-axis and observed values on the y-axis,
different colors correspond to the raw or normalized data. Lines
corresponds to the global median across all samples.
}
\description{
This function can be used to produce a visual aid to for selection
normalization method. Currently \code{\link[mavis]{tmm}} and \code{\link[mavis]{psrn}}
are used. Ideally, after the data is normalized all the boxplots should have
their median aligned close to the global trend line.
}
\examples{
plot_norm_box(yeast_prog, "identifier")
# Plot only ng50 samples
plot_norm_box(yeast_prog, "identifier", plot_target = contains("ng50"))
}
