% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{product}
\alias{product}
\title{Multiply two variables together}
\usage{
product(x, y)
}
\arguments{
\item{x}{Numeric value for one of the variables}

\item{y}{Numeric value of the other variable}
}
\value{
A numeric value of the product of x and y
}
\description{
This function takes the product of x and y
}
\examples{
```{r, child = "man/rmd/child_test.Rmd"}
```

# Unique part
product(x, y)
}
