% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-baldur.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{psrn}
\alias{fit_gamma_regression}
\alias{empirical_bayes}
\alias{weakly_informative}
\alias{infer_data_and_decision_model}
\alias{plot_gamma}
\title{Functions imported from \code{baldur}}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{baldur}{\code{\link[baldur]{empirical_bayes}}, \code{\link[baldur]{fit_gamma_regression}}, \code{\link[baldur]{infer_data_and_decision_model}}, \code{\link[baldur]{plot_gamma}}, \code{\link[baldur]{psrn}}, \code{\link[baldur]{weakly_informative}}}
}}

