% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limma.R
\name{run_limma}
\alias{run_limma}
\title{Run limma and calculate LFC}
\usage{
run_limma(
  data,
  design,
  contrast_matrix,
  gamma_reg_model = NULL,
  id_col = "id",
  weights = NULL,
  .robust = T
)
}
\arguments{
\item{data}{a \code{data.frame} with the samples and feature ids.}

\item{design}{a design or model matrix produced by
\code{\link[stats]{model.matrix}}.}

\item{contrast_matrix}{a contrast matrix produced by
\code{\link[limma]{makeContrasts}}.}

\item{gamma_reg_model}{the regression model produced by
\code{fit_gamma_regression} (see \code{\link[mavis]{fit_gamma_regression}}) or any
\code{glm} with \code{formula} \code{sd ~ ...}.}

\item{id_col}{a character for the name of the column containing the
name of the features in data (e.g., peptides, proteins, etc.).}

\item{weights}{a matrix of precision weights.}

\item{.robust}{Logical value if limma::eBayes should use robust estimator (TRUE) or not (FALSE)}
}
\value{
a \code{tibble} with the id_col, then one p_val_* and lfc_* for each
comparison (*) in the contrast matrix
}
\description{
Function to calculate LFC, run limma using \code{\link[limma]{lmFit}}, \code{\link[limma]{contrasts.fit}},
and \code{\link[limma]{eBayes}} with the flag robust = TRUE.
If neither \code{weights} or \code{gamma_reg_model} is provided, then
\code{\link[limma]{lmFit}} will be run without precision weights and \code{trend = TRUE} (i.e., limma-trend).
If both are provided, then it will default to using the \code{gamma_reg_model}
model to produce the weights.
}
\examples{
# Generate a design matrix for the data
design <- model.matrix(~ 0 + factor(rep(1:2, each = 3)))

# Set correct colnames, this is important for fit_gamma_*
colnames(design) <- paste0("ng", c(50, 100))

# Generate the contrast matrix
contrast <- limma::makeContrasts(
  contrasts = "ng100-ng50",
  levels = design
)

# Normalize and log-transform the data
yeast <- psrn(yeast_prog, "identifier")

# Fit the gamma regressions
gamma_reg_model <- yeast \%>\%
  calculate_mean_sd_trends(design) \%>\%
  fit_gamma_regression(formula = sd ~ mean)

# Exemplify on the non-missing data
yeast <- tidyr::drop_na(yeast)

results <- run_limma(
  yeast,
  design,
  contrast,
  gamma_reg_model,
  "identifier"
)
}
