% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_posterior.R
\name{sample_posterior}
\alias{sample_posterior}
\title{Sample Posterior}
\usage{
sample_posterior(
  data,
  id_col_name,
  design_matrix,
  contrast_matrix,
  uncertainty_matrix,
  bayesian_model = stanmodels$uncertainty_model,
  clusters = 1,
  warmup = 1000,
  iter = 2000,
  chains = 4
)
}
\arguments{
\item{data}{A \code{tibble} or \code{data.frame} with alpha,beta priors annotated}

\item{id_col_name}{A character of the id column}

\item{design_matrix}{A design matrix for the data (see example)}

\item{contrast_matrix}{A contrast matrix of the decisions to test such that the first column is the index of the column in the design matrix that should be compared to the second column}

\item{uncertainty_matrix}{A matrix of observation uncertainty}

\item{bayesian_model}{Which Bayesian model to use. Currently only on internal model allowed, this argument is for forward compatibility}

\item{clusters}{The number of parallel threads to run on.}

\item{warmup}{Number of warm-up samples to draw from the model}

\item{iter}{Total number of iterations to draw (i.e., iter - warmup = posterior draws)}

\item{chains}{The number of chains to sample from}
}
\value{
A \code{tibble} or \code{data.frame} annotated  with statistics of the posterior
}
\description{
Function to sample the posterior of the Bayesian decision model.
Note that, \code{sample_posterior} will remove missing values and adjust the inputs
to the Stan model accordingly. If all values in a condition is missing for a
row/peptide, then that will be removed for that row/peptide.
Still, if there is only one condition that has observed values, \code{sample_posterior}
will crash. Please make sure to impute or filter out these cases.
}
\examples{
# Please see vignette('baldur') for examples
}
