% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MAVIS.R
\name{startaRma}
\alias{startaRma}
\title{Launch aRma interactive User Interface}
\usage{
startaRma()
}
\value{
Nothing
}
\description{
startaRma loads the Turkish version of MAVIS.
}
\details{
Start aRma

This is the Turkish version of MAVIS
}
\examples{
\dontrun{
library(shiny)
startaRma()
}
}
\keyword{aRma}

