% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MAVIS.R
\name{startmavis}
\alias{startmavis}
\title{Launch MAVIS interactive User Interface}
\usage{
startmavis()
}
\value{
Nothing
}
\description{
startmavis starts loads the web browser an interactive user interface built using R shiny.
}
\details{
Start MAVIS

The purpose of building the interactive user interface is to provide an easy for people who are learning how to do their first meta-analysis and/or are not comfortable with the R command line system. Includes example data for testing out a meta-analysis.
}
\examples{
\dontrun{
library(shiny)
startmavis()
}
}
\keyword{MAVIS}

