% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{tmm}
\alias{tmm}
\title{Normalization by Trimmed m Means}
\source{
\url{https://genomebiology.biomedcentral.com/articles/10.1186/gb-2010-11-3-r25}
}
\usage{
tmm(
  data,
  trim_M = 0.3,
  trim_A = 0.05,
  log = TRUE,
  load_info = FALSE,
  target = NULL,
  reference_sample = NULL
)
}
\arguments{
\item{data}{data.frame containing the data to normalize}

\item{trim_M}{percent of fold-change values to trim}

\item{trim_A}{percent of means to trim}

\item{log}{Return log2 transformed values?}

\item{load_info}{Return loading info?}

\item{target}{target columns to normalize, supports
\code{\link[tidyselect]{tidyselect-package}} syntax. By default, all numerical
columns will be used in the normalization if not specified.}

\item{reference_sample}{Specify a reference sample to normalize to, if not
provided, the sample with the lowest coefficient of variation will be used}
}
\value{
data frame with normalized values if \code{load_info=FALSE}, if it is \code{TRUE}
then it returns a list with two tibbles. One tibble containing the
normalized data and one containing the loading info as well as the
estimated normalization factors.
}
\description{
Method for trimmed m means normalization. It is based on the method described
in \insertCite{robinson2010scaling;textual}{mavis}. Though, instead of library
size as was used in the original method, here we use the loading size which
we define as the sum of all features. If no reference sample is specified, it
uses the sample with the lowest coefficient of variation as default.
All estimates are based on features without missing values.
}
\examples{
tmm(yeast)
}
\references{
\insertAllCited{}
}
