% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-baldur.R
\docType{data}
\name{ups}
\alias{ups}
\title{Spiked-in data set of peptides}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10599 rows and 13 columns.
}
\source{
\url{https://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-019-2619-6}
}
\usage{
ups
}
\description{
A dataset containing quantification of peptides using Progenesis. True
positives peptides spiked-in from the Universal Proteomics Standard Set 1
(UPS1) at three different concentrations and true negatives from
\emph{Chlamydomonas reinhardtii} with the same concentration in all samples.
You can find true positives with \code{stringr::str_detect(ups$identifier, 'UPS')}. For details see Berg et al. (2019) and
if you use this dataset please cite the same paper.
}
\references{
{Philip Berg, Evan
W McConnell, Leslie
M Hicks, Sorina
C Popescu, George
V Popescu (2019).
\dQuote{Evaluation of linear models and missing value imputation for the analysis of peptide-centric proteomics.}
\emph{BMC bioinformatics}, \bold{20}(2), 7--16.}{}
}
\keyword{datasets}
