% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-baldur.R
\docType{data}
\name{yeast}
\alias{yeast}
\title{Spiked-in data set of reversibly oxidized cysteines}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2235 rows and 7 columns.
}
\source{
\url{https://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-019-2619-6}
}
\usage{
yeast
}
\description{
A dataset containing quantification of reversibly oxidized cysteines using
Progenesis. True positives cysteines spiked-in from yeast at two different
concentrations and true negatives from \emph{Chlamydomonas reinhardtii} with
the same concentration in all samples. To identify true positives one can use
\code{stringr::str_detect(yeast$identifier, 'YEAST')}. For details see
Berg et al. (2019) and if you use this dataset
please cite the same paper.
}
\references{
{Philip Berg, Evan
W McConnell, Leslie
M Hicks, Sorina
C Popescu, George
V Popescu (2019).
\dQuote{Evaluation of linear models and missing value imputation for the analysis of peptide-centric proteomics.}
\emph{BMC bioinformatics}, \bold{20}(2), 7--16.}{}
}
\keyword{datasets}
