% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{add}
\alias{add}
\title{Add two variables together}
\usage{
add(x, y)
}
\arguments{
\item{x}{Numeric value for one of the variables}

\item{y}{Numeric value of the other variable}
}
\value{
A numeric value of the sum of x and y
}
\description{
This function adds the values of x and y.
}
\examples{
# Common part
x <- 1
y <- 2
# Unique part compared to product
z <- add(x, y)
}
