% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\docType{class}
\name{ensemble}
\alias{ensemble}
\alias{start_ensemble}
\title{Bayesian Ensemble}
\usage{
start_ensemble()
}
\value{
An ensemble object see []
}
\description{
Bayesian Ensemble

Start a New Bayesian Ensemble
}
\examples{
# Setup variables
design <- model.matrix(~ 0 + factor(rep(1:2, each = 3)))
colnames(design) <- paste0("ng", c(50, 100))
contrast <- limma::makeContrasts(
  contrasts = "ng100-ng50",
  levels = design
)

# Exemplify on the non-missing data
yeast <- tidyr::drop_na(yeast) \%>\%
   # Normalize and log-transform the data
   psrn("identifier")

# Fit the gamma regressions
gamma_reg_model <- yeast \%>\%
   calculate_mean_sd_trends(design) \%>\%
   fit_gamma_regression(formula = sd ~ mean)

limma_gr <- run_limma(
  yeast,
  design,
  contrast,
  gamma_reg_model,
  "identifier"
) \%>\%
   head(10)
limma_trend <- run_limma(
  yeast,
  design,
  contrast,
  NULL,
  "identifier"
) \%>\%
   dplyr::filter(identifier \%in\% limma_gr$identifier)

new_ensemble <- start_ensemble()
new_ensemble$add(limma_gr, "GR-limma", "identifier", "p_val", "lfc", do_rm = TRUE
)$add(
   limma_trend, "limma-trend", "identifier", "p_val", "lfc", do_rm = TRUE
)
ensemble_results <- new_ensemble$run_ensemble(iter = 150)
}
