% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{single_imputation}
\alias{single_imputation}
\title{Single imputation}
\usage{
single_imputation(data, design, formula = sd_p ~ mean, workers = 1, ...)
}
\arguments{
\item{data}{a \code{data.frame} to perform the imputation on, missing values should
be \code{NA}.}

\item{design}{a design or model matrix as produced by
\code{\link[stats]{model.matrix}} with column names corresponding to the}

\item{formula}{Regression formula for the pooled variance and mean trend}

\item{workers}{Number of parallel workers to use when building the forest}

\item{...}{Additional arguments to \link[ranger:ranger]{ranger::ranger}.}
}
\value{
a \code{data.frame} with \code{NA} values replaced by imputed values.
}
\description{
Performs a single imputation run and returns the data with NA values replaced
by imputed values.
}
\examples{
# Generate a design matrix for the data
design <- model.matrix(~ 0 + factor(rep(1:2, each = 3)))

# Set correct colnames, this is important for fit_gamma_weights
colnames(design) <- paste0("ng", c(50, 100))

yeast <- yeast_prog \%>\%
  # Normalize and log-transform the data
  psrn("identifier")

# Run the imputation
\dontrun{
yeast \%>\%
  single_imputation(design)
# Run with partitioning
}
}
